Frame-buster
Frame-busting is a jerk move, but so is creating an infinitely nested series of frames. You should now be able to link from any page in our site to any other page in our site and end up at the page you linked to, not as a character on the Droste's cocoa box.